LPS-66598 Move to the right package
