+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_3_2/UpgradeKaleoClassNameAndKaleoClassPKs.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.upgrade.v1_0_0;
+package com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_2;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.StringBundler;
@@ -20,7 +20,6 @@ import com.liferay.portal.workflow.kaleo.model.KaleoNode;
 import com.liferay.portal.workflow.kaleo.model.KaleoTask;
 
 import java.sql.PreparedStatement;
-import java.sql.SQLException;
 
 /**
  * @author Rafael Praxedes
@@ -42,7 +41,11 @@ public class UpgradeKaleoClassNameAndKaleoClassPKs extends UpgradeProcess {
 	protected void upgradeKaleoClassNameAndKaleoClassPK(
 			String tableName, String kaleoClassName,
 			String kaleoClassPKOldColumn)
-		throws SQLException {
+		throws Exception {
+
+		if (!hasColumn(tableName, kaleoClassPKOldColumn)) {
+			return;
+		}
 
 		StringBundler updateScriptSB = new StringBundler(11);
